requirements.txt
flask==2.2.3
python-dotenv==1.0.0
requests==2.28.2
google-generativeai==0.3.1
flask-cors==3.0.10
gunicorn==20.1.0
